HTBHF-1862 Removed hmrc endpoint.
