+++ src/main/java/uk/gov/dhsc/htbhf/smartstub/service/BenefitsService.java
@@ -31,7 +31,6 @@ public class BenefitsService {
     public static final String EXCEPTIONAL_NINO = "ZZ999999D";
     private static final int DWP_ELIGIBILITY_STATUS_FIRST_POSITION = 0;
     private static final int DWP_ELIGIBILITY_STATUS_SECOND_POSITION = 1;
-    private static final int HMRC_ELIGIBILITY_STATUS_POSITION = 1;
     private static final int CHILDREN_UNDER_ONE_POSITION = 2;
     private static final int CHILDREN_UNDER_FOUR_POSITION = 3;
     private static final Map<Character, EligibilityStatus> ELIGIBILITY_STATUS_MAP = Map.of('E', ELIGIBLE, 'I', INELIGIBLE, 'P', PENDING);
@@ -74,35 +73,6 @@ public class BenefitsService {
             return ELIGIBILITY_STATUS_MAP.getOrDefault(ninoChars[DWP_ELIGIBILITY_STATUS_SECOND_POSITION], NO_MATCH);
         }
         return status;
-    }
-
-    public BenefitDTO getHMRCBenefits(String nino) {
-        String householdIdentifier = identifierService.getHMRCHouseholdIdentifier(nino);
-        return getBenefits(nino, HMRC_ELIGIBILITY_STATUS_POSITION, householdIdentifier);
-    }
-
-    private BenefitDTO getBenefits(String nino, int eligibilityStatusPosition, String householdIdentifier) {
-        if (EXCEPTIONAL_NINO.equals(nino)) {
-            String message = "NINO provided (" + EXCEPTIONAL_NINO + ") has been configured to trigger an Exception";
-            log.info(message);
-            throw new IllegalArgumentException(message);
-        }
-        char[] ninoChars = nino.toCharArray();
-        var status = ELIGIBILITY_STATUS_MAP.getOrDefault(ninoChars[eligibilityStatusPosition], NO_MATCH);
-        if (status == NO_MATCH) {
-            return BenefitDTO.builder().eligibilityStatus(NO_MATCH).build();
-        }
-
-        Integer childrenUnderFour = getNumberOfChildrenUnderFour(ninoChars);
-        Integer childrenUnderOne = getNumberOfChildrenUnderOne(childrenUnderFour, ninoChars);
-
-        return BenefitDTO.builder()
-                .eligibilityStatus(status)
-                .numberOfChildrenUnderOne(childrenUnderOne)
-                .numberOfChildrenUnderFour(childrenUnderFour)
-                .householdIdentifier(householdIdentifier)
-                .children(createChildren(childrenUnderOne, childrenUnderFour))
-                .build();
     }
 
     private List<ChildDTO> createChildren(Integer numberOfChildrenUnderOne, Integer numberOfChildrenUnderFour) {
