+++ src/test/java/uk/gov/dhsc/htbhf/smartstub/service/BenefitsServiceTest.java
@@ -13,33 +13,60 @@ import static uk.gov.dhsc.htbhf.eligibility.model.EligibilityStatus.PENDING;
 import static uk.gov.dhsc.htbhf.smartstub.controller.IntegrationTestAssertions.assertCorrectNumberOfChildren;
 import static uk.gov.dhsc.htbhf.smartstub.helper.PersonTestFactory.*;
 
-class BenefitsServiceHMRCTest {
+class BenefitsServiceTest {
 
     private BenefitsService benefitsService = new BenefitsService(new IdentifierService());
 
     @Test
     void shouldReturnIneligibleForMatchingNino() {
-        PersonDTO person = aPersonWhoIsHMRCIneligible();
+        PersonDTO person = aPersonWhoIsDWPIneligible();
 
-        BenefitDTO benefit = benefitsService.getHMRCBenefits(person.getNino());
+        BenefitDTO benefit = benefitsService.getDWPBenefits(person.getNino());
+
+        assertThat(benefit.getEligibilityStatus()).isEqualTo(INELIGIBLE);
+    }
+
+    @Test
+    void shouldReturnIneligibleForAlternateMatchingNino() {
+        PersonDTO person = anAlternatePersonWhoIsDWPIneligible();
+
+        BenefitDTO benefit = benefitsService.getDWPBenefits(person.getNino());
 
         assertThat(benefit.getEligibilityStatus()).isEqualTo(INELIGIBLE);
     }
 
     @Test
     void shouldReturnEligibleForMatchingNino() {
-        PersonDTO person = aPersonWhoIsHMRCEligible();
+        PersonDTO person = aPersonWhoIsDWPEligible();
 
-        BenefitDTO benefit = benefitsService.getHMRCBenefits(person.getNino());
+        BenefitDTO benefit = benefitsService.getDWPBenefits(person.getNino());
+
+        assertThat(benefit.getEligibilityStatus()).isEqualTo(ELIGIBLE);
+    }
+
+    @Test
+    void shouldReturnEligibleForAlternateMatchingNino() {
+        PersonDTO person = anAlternatePersonWhoIsDWPEligible();
+
+        BenefitDTO benefit = benefitsService.getDWPBenefits(person.getNino());
 
         assertThat(benefit.getEligibilityStatus()).isEqualTo(ELIGIBLE);
     }
 
     @Test
     void shouldReturnPendingForMatchingNino() {
-        PersonDTO person = aPersonWhoIsHMRCPending();
+        PersonDTO person = aPersonWhoIsDWPPending();
 
-        BenefitDTO benefit = benefitsService.getHMRCBenefits(person.getNino());
+        BenefitDTO benefit = benefitsService.getDWPBenefits(person.getNino());
+
+        assertThat(benefit.getEligibilityStatus()).isEqualTo(PENDING);
+    }
+
+    @Test
+    void shouldReturnPendingForAlternateMatchingNino() {
+        PersonDTO person = anAlternatePersonWhoIsDWPPending();
+
+        BenefitDTO benefit = benefitsService.getDWPBenefits(person.getNino());
 
         assertThat(benefit.getEligibilityStatus()).isEqualTo(PENDING);
     }
@@ -48,7 +75,7 @@ class BenefitsServiceHMRCTest {
     void shouldReturnNoMatchNino() {
         PersonDTO person = aPersonNotFound();
 
-        BenefitDTO benefit = benefitsService.getHMRCBenefits(person.getNino());
+        BenefitDTO benefit = benefitsService.getDWPBenefits(person.getNino());
 
         assertThat(benefit.getEligibilityStatus()).isEqualTo(NO_MATCH);
         assertThat(benefit.getNumberOfChildrenUnderOne()).isNull();
@@ -59,7 +86,7 @@ class BenefitsServiceHMRCTest {
     void shouldReturnTwoChildrenUnderFourForMatchingNino() {
         PersonDTO person = aPersonWithChildrenUnderFour(2);
 
-        BenefitDTO benefit = benefitsService.getHMRCBenefits(person.getNino());
+        BenefitDTO benefit = benefitsService.getDWPBenefits(person.getNino());
 
         assertThat(benefit.getNumberOfChildrenUnderOne()).isEqualTo(0);
         assertThat(benefit.getNumberOfChildrenUnderFour()).isEqualTo(2);
@@ -70,7 +97,7 @@ class BenefitsServiceHMRCTest {
     void shouldReturnThreeChildrenUnderOneForMatchingNino() {
         PersonDTO person = aPersonWithChildrenUnderOne(3);
 
-        BenefitDTO benefit = benefitsService.getHMRCBenefits(person.getNino());
+        BenefitDTO benefit = benefitsService.getDWPBenefits(person.getNino());
 
         assertThat(benefit.getNumberOfChildrenUnderOne()).isEqualTo(3);
         assertThat(benefit.getNumberOfChildrenUnderFour()).isEqualTo(3);
@@ -81,7 +108,7 @@ class BenefitsServiceHMRCTest {
     void shouldReturnSameNumberOfChildrenUnder1AndUnder4WhenRequestHasUnder1LargerThanUnder4() {
         PersonDTO person = aPersonWithChildren(4, 1);
 
-        BenefitDTO benefit = benefitsService.getHMRCBenefits(person.getNino());
+        BenefitDTO benefit = benefitsService.getDWPBenefits(person.getNino());
 
         assertThat(benefit.getNumberOfChildrenUnderOne()).isEqualTo(1);
         assertThat(benefit.getNumberOfChildrenUnderFour()).isEqualTo(1);
@@ -92,6 +119,6 @@ class BenefitsServiceHMRCTest {
     void shouldThrowExceptionWhenErrorNinoSupplied() {
         PersonDTO person = aPersonWhoWillTriggerAnError();
 
-        assertThrows(IllegalArgumentException.class, () -> benefitsService.getHMRCBenefits(person.getNino()));
+        assertThrows(IllegalArgumentException.class, () -> benefitsService.getDWPBenefits(person.getNino()));
     }
 }
